[refator} VoiceSelection 분리 #179
